+++ packages/storefront/src/components/PartialDocs.vue
@@ -176,10 +176,10 @@ export const viewport: Viewport = {
 };
 
 export const metadata: Metadata = {
-  /* title, description... */
   appleWebApp,
   icons,
-  manifest,
+  /* Next.js currently automatically sets crossorigin="use-credentials" on the manifest link which causes cors problems */
+  /* manifest */
 };`,
         getComponentChunkLinks: `/* ./app/layout.tsx */
 import React from 'react';
@@ -285,6 +285,32 @@ export const Partials = () => {
           {getFontFaceStylesheet({ format: 'jsx' })}
           <!-- Alternative: force using China CDN -->
           {getFontFaceStylesheet({ format: 'jsx', cdn: 'cn' })}
+      </>
+    }
+  });
+
+  return null;
+}
+
+/* render </Partials> component in root layout */`,
+        getFontFaceStyles: `/* ./components/partials.tsx */
+"use client";
+import { useServerInsertedHTML } from "next/navigation";
+import { useRef } from 'react';
+import { getFontFaceStyles } from "@porsche-design-system/components-react/partials";
+
+export const Partials = () => {
+  const isServerInserted = useRef(false);
+
+  useServerInsertedHTML(() => {
+    // There seems to be some strange behavior where Next calls this multiple times
+    // To ensure this is only called once we use a ref
+    if (!isServerInserted.current) {
+      isServerInserted.current = true;
+      return <>
+          {getFontFaceStyles({ format: 'jsx' })}
+          <!-- Alternative: force using China CDN -->
+          {getFontFaceStyles({ format: 'jsx', cdn: 'cn' })}
       </>
     }
   });